requirements.txt
anyio==4.9.0
APScheduler==3.10.1
certifi==2025.4.26
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
typing_extensions==4.13.2
tzlocal==5.3.1
requests==2.31.0  # Для API запросов
beautifulsoup4==4.12.3  # Для парсинга гороскопов
python-telegram-bot[job-queue]==20.3  # Для планировщика